% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{print.recapr_consistencypow}
\alias{print.recapr_consistencypow}
\title{Print method for consistency test power}
\usage{
\method{print}{recapr_consistencypow}(x, ...)
}
\arguments{
\item{x}{Output from \code{powconsistencytest()}}

\item{...}{additional print arguments}
}
\description{
Print method for consistency test power
}
\author{
Matt Tyers
}
