% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{print.recapr_consistencytest}
\alias{print.recapr_consistencytest}
\title{Print method for consistency test}
\usage{
\method{print}{recapr_consistencytest}(x, ...)
}
\arguments{
\item{x}{Output from \code{consistencytest()}}

\item{...}{additional print arguments}
}
\description{
Print method for consistency test
}
\author{
Matt Tyers
}

