% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{print.recapr_stratpow}
\alias{print.recapr_stratpow}
\title{Print method for stratification test power}
\usage{
\method{print}{recapr_stratpow}(x, ...)
}
\arguments{
\item{x}{Output from \code{powstrattest()}}

\item{...}{additional print arguments}
}
\description{
Print method for stratification test power
}
\author{
Matt Tyers
}

