% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{print.recapr_strattest}
\alias{print.recapr_strattest}
\title{Print method for stratification test}
\usage{
\method{print}{recapr_strattest}(x, ...)
}
\arguments{
\item{x}{Output from \code{strattest()}}

\item{...}{additional print arguments}
}
\description{
Print method for stratification test
}
\author{
Matt Tyers
}
