% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdraws_rpvals.R
\name{rBailey}
\alias{rBailey}
\title{Random Draws from the Bailey Estimator}
\usage{
rBailey(length, N, n1, n2)
}
\arguments{
\item{length}{The length of the random vector to return.}

\item{N}{The value of the true abundance. This may be a single number or
vector of values.}

\item{n1}{Number of individuals captured and marked in the first sample. This
may be a single number or vector of values.}

\item{n2}{Number of individuals captured in the second sample.  This may be a
single number or vector of values.}
}
\value{
A vector of random draws from the Bailey estimator
}
\description{
Returns a vector of random draws from the Bailey estimator in a
  mark-recapture experiment, given values of the true abundance and the
  sample size in both events.  The function first simulates a vector of
  recaptures (m2) from a binomial distribution, and then uses these to
  compute a vector of draws from the estimator.
}
\note{
Any Petersen-type estimator (such as this) depends on a set of
  assumptions: \itemize{ \item  The population is closed; that is, that there
  are no births, deaths, immigration, or emigration between sampling events
  \item All individuals have the same probability of capture in one of the
  two events, or complete mixing occurs between events \item Marking in the
  first event does not affect probability of recapture in the second event
  \item Individuals do not lose marks between events \item All marks will be
  reported in the second event }
}
\examples{
draws <- rBailey(length=100000, N=500, n1=100, n2=100)
plotdiscdensity(draws)  #plots the density of a vector of discrete values
}
\author{
Matt Tyers
}
\seealso{
\link{NBailey}, \link{vBailey}, \link{seBailey}, \link{pBailey},
  \link{powBailey}, \link{ciBailey}
}

