\name{recapr-package}
\alias{recapr-package}
\alias{recapr}
\docType{package}
\title{
Estimating, Testing, and Simulating Abundance in a Mark-Recapture
}
\description{
Tools are provided for estimating, testing, and simulating abundance in a two-event (Petersen) mark-recapture experiment.  Functions are given to calculate the Petersen, Chapman, and Bailey estimators and associated variances.  However, the principal utility is a set of functions to simulate random draws from these estimators, and use these to conduct hypothesis tests and power calculations.  Additionally, a set of functions are provided for generating confidence intervals via bootstrapping.  Functions are also provided to test abundance estimator consistency under complete or partial stratification, and to calculate stratified or Darroch estimators.  Functions are also provided to calculate recommended sample sizes.
}
\details{
\tabular{ll}{
Package: \tab recapr\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2016-05-30\cr
License: \tab GPL-2\cr
}
}
\author{
Matt Tyers

Maintainer: Matt Tyers <matt.tyers@alaska.gov>
}


