% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{sePetersen}
\alias{sePetersen}
\title{Standard Error of the Petersen Estimator}
\usage{
sePetersen(n1, n2, m2)
}
\arguments{
\item{n1}{Number of individuals captured and marked in the first sample. This
may be a single number or vector of values.}

\item{n2}{Number of individuals captured in the second sample.  This may be a
single number or vector of values.}

\item{m2}{Number of marked individuals recaptured in the second sample.  This
may be a single number or vector of values.}
}
\value{
The estimate variance of the Petersen estimator, calculated as
  sqrt((n1^2)*n2*(n2-m2)/(m2^3))
}
\description{
Calculates the standard error of the Petersen estimator in a
  mark-recapture experiment, with given values of sample sizes and number of
  recaptures.
}
\note{
Any Petersen-type estimator (such as this) depends on a set of
  assumptions: \itemize{ \item  The population is closed; that is, that there
  are no births, deaths, immigration, or emigration between sampling events
  \item All individuals have the same probability of capture in one of the
  two events, or complete mixing occurs between events \item Marking in the
  first event does not affect probability of recapture in the second event
  \item Individuals do not lose marks between events \item All marks will be
  reported in the second event }
}
\examples{
sePetersen(n1=100, n2=100, m2=20)
}
\seealso{
\link{NPetersen}, \link{vPetersen}, \link{rPetersen}, \link{pPetersen},
  \link{powPetersen}, \link{ciPetersen}
}
\author{
Matt Tyers
}
